requirements.txt
torch>=2.0
torchvision>=0.15
Pillow>=10.0
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
ultralytics>=8.0
